+++ src/App.js
@@ -1,18 +1,13 @@
 import React from 'react';
-import Button from './stories/StoryButton';
+// import Button from './stories/StoryButton';
+import Button from './components/Button';
 import './global.css';
 
 function App() {
 	return (
 		<div className="App">
 			<h1>Moyo</h1>
-			<Button
-				backgroundColor={null}
-				label="Moyo"
-				onClick={() => {}}
-				size="large"
-				primary
-			/>
+			<Button label="Moyo" onClick={() => {}} size="large" primary />
 		</div>
 	);
 }
