+++ src/components/Button.css
@@ -1,0 +1,24 @@
+.moyo-button {
+    color: white;
+    font-weight: 700;
+	border-radius: 20px;
+	background: var(--moyo-green-400);
+	cursor: pointer;
+	padding: 10px 16px;
+	outline: none;
+	border: none;
+}
+
+.moyo-button--medium {
+	font-size: 14px;
+	padding: 11px 20px;
+}
+
+.moyo-button--large {
+	font-size: 16px;
+	padding: 12px 24px;
+}
+
+.moyo-button:active {
+	transform: scale(0.95);
+}
