+++ src/stories/Button.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 const Button = ({ label, size, ...props }) => {
 	return (
 		<button
-			className={['moyo-button', `moyo-button-${size}`]}
+			className={['moyo-button', `moyo-button--${size}`]}
 			{...props}
 		>
 			{label}
